requirements.txt
cloudinary==1.39.1
Flask==3.0.3
flask_bcrypt==1.0.1
flask_login==0.6.3
flask_mail==0.9.1
flask_sqlalchemy==3.1.1
itsdangerous==2.2.0
psycopg2==2.9.9
python-dotenv==1.0.1
pytz==2024.1
Requests==2.31.0
SQLAlchemy==2.0.29
Werkzeug==3.0.2
gunicorn==20.1.0